% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots.R
\name{as_dots}
\alias{as_dots}
\title{convert object to dots}
\usage{
as_dots(x)
}
\arguments{
\item{x}{a list-like object, including a list of quosures. In the former case
the environments will be set to the caller environment, in the latter
case the environments will be preserved.}
}
\description{
convert object to dots
}
