% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{dot-subset}
\alias{dot-subset}
\alias{[.dots}
\alias{[[.dots}
\alias{$.dots}
\title{Subset dots}
\usage{
\method{[}{dots}(x, i, ...)

\method{[[}{dots}(x, i, ...)

\method{$}{dots}(x, name, ...)
}
\arguments{
\item{x}{a "dots" object}

\item{i}{an index for \code{[} or \code{[[}}

\item{...}{ignored}

\item{name}{a character index for \code{$}}
}
\description{
Use \code{[} to subset or reorder dots, this will not trigger any evaluation.
Use \code{[[} or \code{$} to evaluate an argument, note that \code{x} will not be modified.
}
