% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots.R
\name{dot_split}
\alias{dot_split}
\title{split a "dots" object into single argument "dots"}
\usage{
dot_split(x, ...)
}
\arguments{
\item{x}{A "dots" object built with \code{dots()}}

\item{...}{ignored}
}
\value{
A list of single argument "dots" objects
}
\description{
split a "dots" object into single argument "dots"
}
\examples{
f <- function(...) {
  dots1 <- dots()
  dot_split(dots1)
}
f(a = x, y)
}
