% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots.R
\name{dots_env}
\alias{dots_env}
\alias{dots_env<-}
\title{Get or Set the Environment of a 'dots' Object}
\usage{
dots_env(x)

dots_env(x) <- value
}
\arguments{
\item{x}{a 'dots' object}

\item{value}{an environment}
}
\description{
Get or Set the Environment of a 'dots' Object
}
