% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots.R
\name{enarg}
\alias{enarg}
\title{transform a one argument "dots" object into an argument}
\usage{
enarg(x, nm = NULL, ...)
}
\arguments{
\item{x}{a one argument "dots" object}

\item{nm}{an optional variable name, if absent the name provided to the dots
will be used}

\item{...}{unused}
}
\description{
transform a one argument "dots" object into an argument
}
\examples{
f <- function(...) {
  d <- dots(...)
  enarg(d)
  enarg(d, "y")
  message("not yet evaled")
  x + y
}
f(x = 1)
try(f(x = stop()))
}
