% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots.R
\name{endot}
\alias{endot}
\title{transform an argument into a one argument "dots" object}
\usage{
endot(x)
}
\arguments{
\item{x}{an argument}
}
\description{
transform an argument into a one argument "dots" object
}
\examples{
f <- function(x) {
  dots1 <- endot(x)
  dots1
}
f(y+z)
}
