% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots.R
\name{with_dots}
\alias{with_dots}
\title{Evaluate an Expression with dots}
\usage{
with_dots(dots, expr)
}
\arguments{
\item{dots}{a "dots" object}

\item{expr}{an expression using \code{...}}
}
\description{
\code{with_dots(d, f(...))} is equivalent to \code{with(list(... = d), f(...))} but
shorter and quicker.
}
\examples{
dots1 <- dots(letters, n =3)
dots1
with_dots(dots1, head(...))
dots2 <- dots(n =3)
dots2
with_dots(dots2, head(letters, ...))
}
